projeto_risco_cardiaco/projeto_risco_cardiaco/requirements.txt
# Kaggle API client (download datasets)
kaggle>=1.6.0

# Arrow / Parquet (high-performance CSV reading + Parquet writing)
pyarrow>=14.0.0

# YAML config parsing
PyYAML>=6.0.1

# .env file loading (secrets.env support)
python-dotenv>=1.0.0

# ── EDA Core ───────────────────────────────────────────────────────────────────
# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Statistical testing (ANOVA, Kruskal-Wallis, normality tests, etc.)
scipy>=1.11.0

# Post-hoc tests (Tukey HSD) and advanced stats
statsmodels>=0.14.0

# Machine learning (K-Means, DBSCAN, PCA, StandardScaler, silhouette)
scikit-learn>=1.3.0

# ── EDA Visualization ─────────────────────────────────────────────────────────
# Core plotting
matplotlib>=3.7.0

# Statistical visualization (violin plots, KDE, pairplots, heatmaps)
seaborn>=0.12.0

# ── Data Quality ──────────────────────────────────────────────────────────────
# Framework de validação de dados com expectations declarativas (via YAML)
# great-expectations>=1.0.0

# ── Modelagem e Experimentação ────────────────────────────────────────────────
# Rastreamento de experimentos, métricas, artefatos e model registry
# mlflow>=2.10.0

# Otimização de hiperparâmetros (Bayesian/TPE, pruning, visualizações)
# optuna>=3.5.0

# Gradient boosting otimizado (histogramas, GPU-ready)
xgboost>=2.0.0

# Gradient boosting baseado em histogramas — mais rápido que XGBoost em datasets grandes
lightgbm>=4.2.0